Moving Positive and Relaxed to containers, refactor a bit to use async await, putting some extra css to cater removal of react materialize
